update deploy script & add ci
