+++ scripts/utils.ts
@@ -1,7 +1,7 @@
 import { Contract } from "ethers";
 import hre, { ethers, upgrades } from "hardhat";
 
-async function deployAndVerify({
+export async function deploy({
   name,
   args,
   isProxy,
@@ -18,7 +18,21 @@ async function deployAndVerify({
     contract = await ethers.deployContract(name, args);
   }
 
-  console.log(`${name} contract is deployed to ${contract.target}`);
+  console.log(`${name} contract is deployed to ${await contract.getAddress()}`);
+
+  return contract;
+}
+
+export async function deployAndVerify({
+  name,
+  args,
+  isProxy,
+}: {
+  name: string;
+  args: any[];
+  isProxy: boolean;
+}) {
+  const contract = await deploy({ name, args, isProxy });
 
   // npx hardhat verify [CONTRACT_ADDRESS] [...CONSTRUCTOR_ARGS] --network alfajores
   if (contract.deploymentTransaction()?.hash) {
@@ -29,33 +43,3 @@ async function deployAndVerify({
     });
   }
 }
-async function main() {
-  const contracts = [
-    {
-      name: "CRecyTest",
-      args: [],
-      isProxy: false,
-    },
-    {
-      name: "FCWERC721",
-      args: [],
-      isProxy: false,
-    },
-    {
-      name: "TimeLock",
-      args: ["0xf64C1a07144B22cdD109d5b82004aEd4759114c4"],
-      isProxy: true,
-    },
-  ];
-
-  for (const contract of contracts) {
-    await deployAndVerify(contract);
-  }
-}
-
-// We recommend this pattern to be able to use async/await everywhere
-// and properly handle errors.
-main().catch((error) => {
-  console.error(error);
-  process.exitCode = 1;
-});
